+++ backend-python/app/repositories/appointment_repository.py
@@ -2,8 +2,10 @@ from pymongo import MongoClient
 from bson.objectid import ObjectId
 from dotenv import load_dotenv
 import os
+from exceptions import AppointmentNotFoundException, ValidationError, DatabaseError
 
 load_dotenv()
+
 
 class AppointmentRepository:
     def __init__(self):
@@ -12,17 +14,34 @@ class AppointmentRepository:
         self.collection = self.db[os.getenv('APPOINTMENTS_COLLECTION')]
 
     def create_appointment(self, data):
-        result = self.collection.insert_one(data)
-        return str(result.inserted_id)
+        if not data.get('patient_id') or not data.get('doctor_id'):
+            raise ValidationError("Appointment must have a patient_id and a doctor_id")
+
+        try:
+            result = self.collection.insert_one(data)
+            inserted_document = self.collection.find_one({"_id": result.inserted_id})
+            return inserted_document
+        except Exception as e:
+            raise DatabaseError(f"Error creating appointment: {str(e)}")
 
     def get_appointment_by_id(self, appointment_id):
-        return self.collection.find_one({'_id': ObjectId(appointment_id)})
+        appointment = self.collection.find_one({'_id': ObjectId(appointment_id)})
+        if not appointment:
+            raise AppointmentNotFoundException(appointment_id)
+        return appointment
 
     def get_all_appointments(self):
-        return list(self.collection.find({}))
+        try:
+            return list(self.collection.find({}))
+        except Exception as e:
+            raise DatabaseError(f"Error fetching appointments: {str(e)}")
 
     def update_appointment(self, appointment_id, data):
+        if not self.get_appointment_by_id(appointment_id):
+            raise AppointmentNotFoundException(appointment_id)
         self.collection.update_one({'_id': ObjectId(appointment_id)}, {'$set': data})
 
     def delete_appointment(self, appointment_id):
+        if not self.get_appointment_by_id(appointment_id):
+            raise AppointmentNotFoundException(appointment_id)
         self.collection.delete_one({'_id': ObjectId(appointment_id)})
